+++ systems-client/src/main/java/edu/utexas/tacc/tapis/systems/client/SystemsClient.java
@@ -269,7 +269,7 @@ public class SystemsClient
 
   /**
    * Get list of systems using all supported parameters: searchStr, limit, sortBy, sip, startAfter.
-   * For example search=(id.like.MySys*)~(enabled.eq.true)&limit=10&sortBy=seqId(asc)&startAfter=101
+   * For example search=(id.like.MySys*)~(enabled.eq.true)&limit=10&sortBy=id(asc)&startAfter=my.sys1
    * Use only one of skip or startAfter
    * When using startAfter sortBy must be specified.
    */
@@ -305,7 +305,7 @@ public class SystemsClient
    * Dedicated search endpoint using all supported parameters
    * Search for systems using an array of strings that represent an SQL-like WHERE clause
    * and using query parameters for sorting.
-   * For example limit=10&sortBy=seqId(asc)&startAfter=101
+   * For example limit=10&sortBy=id(asc)&startAfter=my.sys1
    * Use only one of skip or startAfter
    * When using startAfter sortBy must be specified.
    */
@@ -352,13 +352,14 @@ public class SystemsClient
    * Return 0 if record not present
    *
    * @param systemId System systemId
+   * @param confirm Confirm the action
    * @return number of records modified as a result of the action
    * @throws TapisClientException - If api call throws an exception
    */
-  public int deleteSystem(String systemId) throws TapisClientException
+  public int deleteSystem(String systemId, Boolean confirm) throws TapisClientException
   {
     RespChangeCount resp = null;
-    try { resp = sysApi.deleteSystem(systemId, false); }
+    try { resp = sysApi.deleteSystem(systemId, false, confirm); }
     catch (ApiException e) { Utils.throwTapisClientException(e.getCode(), e.getResponseBody(), e); }
     catch (Exception e) { Utils.throwTapisClientException(-1, null, e); }
     if (resp != null && resp.getResult() != null && resp.getResult().getChanges() != null) return resp.getResult().getChanges();
